Update README to list java 8 as requirement
Because I just had to play with the java.time library and remember that
I updated my local machine to run java 8... which means I could probably
try out a later version of play if I wanted to.

But do I want to? Eh, I enjoy 2.3 and am comfortable with it. Maybe I'll
code the entire application in 2.3 then see how much fun it is to follow
the migration guide to 2.4 or something.