Add ExpenseTag class
Not sure if I'm going to call this ExpenseTag or ExpenseGroup yet. It's
going to be the initial replacement for 'accounts' in BGI. Which means
that we only need to support an initial 1-M relationship, but when I
think about invoicing, it would make sense to call it a group. Plus, a
group _feels_ like it implies a one group many expenses relationship,
while a tag doesn't since the language implies (to me) that an expense
could have many tags.