Remove ExpenseTag, Create ExpenseGroup
I basically convinced myself in my last commit message that semantically
I'd find it easier to work with an ExpenseGroup rather than with
ExpenseTags at first. I think they're a more natural fit for BGI's
"accounts".

I think the next step will be to bootstrap a simple service for creating
and reading Expenses. Then I'll create a service to create and read
groups with. Then I'll update the services to be able to add expenses to
groups. From there I'll be able to work on an import service.