Add test conf file and update build for it
Updating build to point to the new application.conf file in the test
folder. This will include the other one, then modify only what it needs
to in order to switch the connection to the test database.

Though it would be really nice to set things up so that it creates a
fake test database and then tears it down after. that might be a fun
exercise later on...