Add expenseGroup parser and joined parser
Have to have the joined parser because trying to use both the expense
and expensegroup parser combined with ~ seems to give issues trying to
resolve the column name with the case statements.

If I can figure out how to get anorm to understand the difference, or if
it can handle aliasing a column as the name from the table then I can
drop that one, but for now I need to do this I guess.