Add DataTooLargeException
Creating an application exception for handling database column size
constraint violations.

It would be nice to figure out a way to reliably determine _what_ field
was the issue and not just which index in the parameter list it is, but
doesn't seem doable with the exception type this is going to be
catching.

On the other hand, hopefully we'll never see this anyway because
validation in the upper levels such as controller, form, and service
should help us stear clear of that.