Start in on ExpenseManagementService
Not going to make two seperate services for the expense and expense
group because it seems easy to get confused on remembering where a
createExpeneAndAddToGroup method would live, and so I'll just put them
all into one management service.

The service layer will bulk up once I start adding in other pieces of
funtionality, like the actual accountability part of the application and
the email notifications.