Fix bug in ExpenseManagementService
Was not properly waiting for the two futures. While I thought I was
chaining properly by mapping/recovering one future into the other, I was
not transforming the first future into the second one. Which I need
flatMap for. The key thing here is in the scaladoc:

	FlatMap:
	Creates a new future by applying a function to the successful result
	of this future, and returns the result of the function as the new
	future. If this future is completed with an exception then the new
	future will also contain this exception.

Plain map is missing the "and returns the result of the function as the
new future" part.