Add high level todo list
Just to keep an idea of what I want to do with this and finish it up.
The emailing to add expenses has been an idea of mine for a while, and
I'm not sure how hard it is to do, but I know it's possible.

Another thing instead of that could be to maybe dive into android again
and make an app for it. That's a big maybe though and I'm unlikely to do
that I think since last time I programmed for android it was a PITA