Flush out create expense action in ExpenseController
Still need to actually create some viewss and such, but the basic
functionality is encoded for creating an expense even if the way the
final results, error or success, are shown to the user are currently
sup-par.

Need to skeleton an action for displaying a create expense form, and
then I think I'll be good to work on that small part of the view.

Once that's done then I can do the expense group's forms and whatnot.
Which should be pretty easy.