Add some views and imports for templates
Update ExpenseController to call out to the currentExpenses template,
made a couple common helpers as well. Havent gotten around to figuring
out how I'm going to style this yet, but I'll figure it out at some
point.

Also imported the common models into the templates because they'll be
the arguments to most templates in all likelihood