Add TemplateNotifications class
Even though play has a the flash scope to pass arbitrary messages out to
the template, I'm going to use this class to do most of that. Reason
being that in order to use .flashing, you need to redirect and I don't
always intend on redirecting like that when I might need to display an
error or something to the user.