Add NotifyingController trait
Going to use an implicit to provide the templateNotifications to the
partials, want to make this be as little boilerplate as possible, so
going to create an implicit converstion of Throwable ->
List[TemplateNotification] so that we can convert throwables to the
templates easily.

Sadly, just having a Throwable in scope doesn't convert it since we need
the list of notifications to be implicit,  so I made a helper method
`withErrorPage` that does it more explicitly.