Add some table style
Definitely not the final color scheme obviously, but just trying to get
a feel for some variations and places where I will color in the tables a
bit more once I figure out an actual color-scheme for this thing.

I used green mainly in BGI, so I don't know if I'll go with that again
or maybe make a darker-theme so that I can throw it up when I'm about to
go to sleep and want to darken my screen a bit. Hm.