Create manageGroups template skeleton
Make the controller method to list the groups
Update the navigation to point to the new controller route
Create the template for listing the groups

The template isn't done yet, we still need to make a new of other style
changes to it, and add in the actions we'll be able to do. But that will
come soon.