Show groups without expenses for a week in summary
Before, the summary page would only show the expense groups for a given
week if there was an expense in that category for that week. But, in
this kind of table listing, I'd rather see all groups listed even if I
didn't have any expenses in that week for it.