Able to switch groups of a created expense
It's not pretty looking though; bloats up the table quite a bit so I
really don't want to keep it the way it is. Might be better to either
use a seperate page for this and just link to it from here, or create
one form on this page that can take data from the table and populate the
form as neccesary with it.