Add select week page and supporting methods
To select a previous week to see a summary. I need to make the actual
summary route this will link to, but the rough CSS and display of the
old expenses and whatnot is done for now.

I might later on add some cumulative information, like maybe a summary
of all expenses to go with each week. But for now this is fine.